Fix clang compilation errors (signdness on cast)

+++ src/lib/mdns/minimal/BytesRange.h
@@ -42,7 +42,7 @@ public:
 
     bool Contains(const uint8_t * p) const { return ((p >= mStart) && (p < mEnd)); }
 
-    size_t Size() const { return (mEnd - mStart); }
+    size_t Size() const { return static_cast<size_t>(mEnd - mStart); }
 
 private:
     const uint8_t * mStart = nullptr;
